test_vocal_extraction/requirements.txt
# Core dependencies
numpy==1.23.5
scipy==1.15.3
librosa==0.10.1
soundfile==0.12.1
matplotlib==3.8.0

# For Demucs
torch==2.3.0
demucs==4.0.1

# For Spleeter (requires Python 3.11 with arm64 architecture)
tensorflow-macos==2.12.0
tensorflow-metal==1.2.0
spleeter==2.4.2
numba==0.57.1
httpx==0.28.1
h2==4.2.0
pandas==2.3.1
